Move to helpers/remoteoauth
